move ffi_client.h to internal src/ folder
