+++ src/audio_source.cpp
@@ -22,8 +22,8 @@
 
 #include "audio_frame.pb.h"
 #include "ffi.pb.h"
+#include "ffi_client.h"
 #include "livekit/audio_frame.h"
-#include "livekit/ffi_client.h"
 
 namespace livekit {
