+++ src/remote_audio_track.cpp
@@ -17,8 +17,8 @@
 #include "livekit/remote_audio_track.h"
 
 #include "ffi.pb.h"
+#include "ffi_client.h"
 #include "livekit/audio_source.h"
-#include "livekit/ffi_client.h"
 #include "track.pb.h"
 #include "track_proto_converter.h"
