+++ src/video_source.cpp
@@ -19,10 +19,9 @@
 #include <chrono>
 #include <stdexcept>
 
-#include "livekit/ffi_client.h"
-#include "livekit/video_frame.h"
-
 #include "ffi.pb.h"
+#include "ffi_client.h"
+#include "livekit/video_frame.h"
 #include "video_frame.pb.h"
 #include "video_utils.h"
