+++ src/video_utils.cpp
@@ -5,7 +5,7 @@
 #include <vector>
 
 #include "ffi.pb.h"
-#include "livekit/ffi_client.h"
+#include "ffi_client.h"
 #include "livekit/ffi_handle.h"
 #include "video_frame.pb.h"
